Separate shared SQL adapter behaviour specs
