+++ spec/statesman/adapters/adapter_behaviour.rb
