+++ spec/statesman/adapters/mongoid_spec.rb
@@ -1,5 +1,5 @@
 require "spec_helper"
-require "statesman/adapters/shared_examples"
+require "statesman/adapters/adapter_behaviour"
 require "statesman/exceptions"
 require "support/mongoid"
 require "mongoid"
